+++ config/.env.base.sh
@@ -15,6 +15,7 @@ export YLPZ_RA_LAD_001_NL_REV=0
 export YLPZ_RA_LAD_001_IT_REV=0
 export YLPZ_RA_LAD_001_ES_REV=0
 export YLPZ_RA_LAD_001_DE_REV=0
+export IDLE_TIMEOUT_MS=1800000
 
 # External services:
 # URL to HELP system