+++ packages/yourloops/lib/auth/models/authenticated-user.model.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, Diabeloop
+ * Copyright (c) 2022-2023, Diabeloop
  *
  * All rights reserved.
  *
@@ -34,3 +34,5 @@ export interface AuthenticatedUser {
   sub: string
   frProId?: string
 }
+
+export const IDLE_USER_QUERY_PARAM = 'idle'