+++ packages/yourloops/test/unit/lib/auth/auth.hook.test.tsx
@@ -59,7 +59,7 @@ describe('Auth hook', () => {
     email: 'fake@email.com',
     hcpProfession: HcpProfession.diabeto
   }
-  const preferences: Preferences = { displayLanguageCode: 'en' }
+  const preferences: Preferences = { displayLanguageCode: LanguageCodes.En }
   const settings: Settings = { country: CountryCodes.France, units: { bg: UnitsType.MGDL } }
 
   const initAuthContext = async (): Promise<void> => {
@@ -112,6 +112,16 @@ describe('Auth hook', () => {
       expect(auth.isLoggedIn).toBeTruthy()
       await auth.logout()
       expect(auth0Mock.useAuth0().logout).toHaveBeenCalledTimes(1)
+      expect(auth0Mock.useAuth0().logout).toHaveBeenCalledWith({ returnTo: 'http://localhost/login' })
+    })
+
+    it('should logout the logged-in idle user', async () => {
+      await initAuthContext()
+      expect(auth.user).not.toBeNull()
+      expect(auth.isLoggedIn).toBeTruthy()
+      await auth.logout(true)
+      expect(auth0Mock.useAuth0().logout).toHaveBeenCalledTimes(1)
+      expect(auth0Mock.useAuth0().logout).toHaveBeenCalledWith({ returnTo: 'http://localhost/login?idle=true' })
     })
   })